Word Generator..
A simple word generator. In this word generator, if you click any alphabet, it will generate a word starting with that letter. If you type any letter with your keyboard, it will work the same way. Additionally, if you click any letter, the word will be displayed, and the program will provide information about the word.